+++ NEWS.md
@@ -1,5 +1,6 @@
 # jtools 2.2.2
 
+* Corrected a S3 methods argument mismatch on the internal function `test_coefs()`.
 * In addition the point size, shape, and color, you may now also change the 
 thickness of the lines drawn to represent coefficient intervals with the 
 `plot_summs()` and `plot_coefs()` functions. You need only use the `line.size`